Update README
added idea of "prod" branch and keeping "master" sacred.
Not sure if this branch scheme is actually gonna work
like what if i clone a linux system, then push to the linux branch
I want those updates in master as well, so PR it?
but will the PR pull in the commit that removed the Windows and
Examples?

So many questions. So little time

maxprehl